+++ src/components/CopyUiProvider/index.ts
@@ -1,5 +1,6 @@
-// CopyUiProvider from copy-ui @ 2025-11-08.
+// CopyUiProvider from copy-ui @ 2025-11-15.
 
 export { CopyUiProvider } from "./copy-ui-provider";
+export type { ColorName, ColorNo } from "./theme";
 export { useJss } from "./use-jss";
 export { useTheme } from "./use-theme";